V2: New folder structure + redux modules
