+++ src/common/components/App/App.scss
@@ -1,0 +1,4 @@
+.container {
+  padding: 20px;
+  border: solid 1px #212121;
+}
