+++ src/common/components/PostListItem/index.js
@@ -2,18 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 // import { Link } from 'react-router-dom';
 
-// Import Style
-// import styles from './PostListItem.scss';
-const styles = [];
-
 function PostListItem(props) {
   return (
-    <div className={styles['single-post']}>
-      <h3 className={styles['post-title']}>
+    <div>
+      <h3>
         {props.post.title}
       </h3>
-      <p className={styles['post-desc']}>{props.post.content}</p>
-      <hr className={styles.divider} />
+      <p>{props.post.content}</p>
+      <hr />
     </div>
   );
 }
