--- src/common/modules/Post/components/PostList/PostList.scss
@@ -1,39 +1,0 @@
-.CartList {
-  padding: 0;
-  .CartList-item {
-    width: 100%;
-    background-color: #E2E2E2;
-    padding: 15px;
-    list-style: none;
-    border: solid 1px #CCC;
-    border-left: 0;
-    border-right: 0;
-    border-bottom: 0;
-    .CartList-item-remove {
-      float: right;
-      color: #444;
-      cursor: pointer;
-      &:hover {
-        text-decoration: underline;
-      }
-    }
-    .CartList-item-price {
-      margin-left: 20px;
-    }
-    &:nth-child(2n+2) {
-      background-color: #EEE;
-    }
-    .CartList-amount {
-      float: right;
-      width: 30px;
-      text-align: center;
-      margin-right: 15px;
-    }
-  }
-  li:last-child {
-    list-style: none;
-    text-align: right;
-    width: 100%;
-    margin: 10px;
-  }
-}
