--- src/common/modules/Post/components/PostListIndex.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import Helmet from 'react-helmet';
-
-const PostListIndex = () => (
-  <div>
-    <Helmet title="Post List" />
-    <h1>PostListIndex</h1>
-  </div>
-);
-
-PostListIndex.displayName = 'PostListIndex';
-
-export default PostListIndex;
