--- src/common/modules/Post/components/PostListItem/PostListItem.scss
@@ -1,58 +1,0 @@
-.single-post {
-  margin: 20px 0;
-  padding: 15px;
-  border-radius: 2px;
-}
-
-.post-title{
-  font-size: 28px;
-  margin-bottom: 16px;
-  font-weight: 400;
-  color: #616161;
-}
-
-.post-title a{
-  text-decoration: none;
-  color: #616161;
-}
-
-.author-name{
-  font-size: 16px;
-  margin-bottom: 16px;
-  color: #757575;
-}
-
-.post-desc{
-  font-size: 14px;
-  color: #888;
-  margin-bottom: 8px;
-}
-
-.post-action a{
-  color: #555;
-  text-decoration: none;
-  font-size: 14px;
-  font-style: italic;
-}
-
-.post-action a:hover{
-  color: #EF5350;
-}
-
-.divider{
-  border: 0;
-  height: 1px;
-  background: #ccc;
-  width: 250px;
-  margin: 32px auto 0;
-}
-
-.post-detail .post-title{
-  font-size: 42px;
-  color: #454545;
-}
-
-.post-detail .post-desc{
-  font-size: 16px;
-  color: #555;
-}
