--- src/common/modules/Post/pages/PostListPage/PostListPage.scss
@@ -1,11 +1,0 @@
-.PostPage {
-  .title {
-    position: relative;
-    padding: 15px;
-    background-color: #f6f8fa;
-    border: 1px solid rgba(27, 31, 35, 0.15);
-    border-radius: 3px 3px 0 0;
-    font-size: 18px;
-    font-family: monospace;
-  }
-}
