+++ src/common/redux/utils/reducer.js
@@ -1,0 +1,4 @@
+export default (initialState, actionHandlers, action) => {
+  const handler = actionHandlers[action.type];
+  return handler ? handler(initialState, action) : initialState;
+};
